# src/model/dto/setupdto/setup_request_dto.ts
export interface SetupRequestDto {
    codigo: number;
    nome_conta: string;
    grupo_contabil: string;
    subgrupo1?: string | null;
    subgrupo2?: string | null;
    saldo_inicial?: number | null;
}